Centralize app init in main script
The initialization of the app has been made much cleaner and saner. All
the important things happen before the initial render and even before
the setup of the store.

First Ethereum and IPFS go online
Then the accounts are fetched and a primary account is set
Then the Redux store is setup with initial state so far
All the relevant data is fetched like messages of primary account
And the first render happens.

In the future, all of the failures that could happen above before the
first render will be handled accordingly.